Finish migration to the Filesystem interface.
